group update, create and style

+++ backend/src/groups/groups.service.ts
@@ -3,10 +3,10 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Group } from './groups.entity';
 import { GroupMembersService } from 'src/group-members/group-members.service';
-import { UsersService } from 'src/users/users.service';
 import { CreateGroupDto } from './dto/create-group.dto';
 import { UpdateGroupDto } from './dto/update-group.dto';
-import { log } from 'console';
+import { User } from 'src/users/users.entity';
+import { Role } from 'src/enums/role.enum';
 
 @Injectable()
 export class GroupsService {
@@ -64,8 +64,7 @@ export class GroupsService {
   }
 
   async updateGroup(updateGroupDto: UpdateGroupDto): Promise<Group> {
-    const { groupId, groupName, groupPicture, gameTypes, userIds } =
-      updateGroupDto;
+    const { groupId, groupName, groupPicture, gameTypes } = updateGroupDto;
 
     const group = await this.findGroupById(groupId);
 
@@ -78,14 +77,20 @@ export class GroupsService {
     if (gameTypes !== undefined) {
       group.gameTypes = [...gameTypes];
     }
-    if (userIds !== undefined) {
-      await this.groupMemberService.addUsersToGroup(groupId, userIds);
-    }
+
     return await this.groupRepository.save(group);
   }
 
-  async deleteGroup(groupId: string): Promise<void> {
+  async deleteGroup(groupId: string, user: User): Promise<void> {
     const group = await this.findGroupById(groupId);
+    const isManager = group.groupMembers.some(
+      (member) => member.user.uid === user.uid && member.isManager,
+    );
+
+    if (!isManager && user.role !== Role.ADMIN) {
+      throw new NotFoundException('אין לך הרשאות למחוק קבוצה זו');
+    }
+
     await this.groupRepository.remove(group);
   }
 }
